Add main.go file with code to fetch and print Pokemon name and nature
This commit adds a new file called main.go which contains code to fetch and print the name of a Pokemon (in this case, Pikachu) using the PokeAPI library. It also fetches and prints the name of a nature (in this case, Adamant). If there are any errors during the API calls, they will be printed as well.